# flyio-admin/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "flyio-admin API",
    "description": "Application and Orchestration layer API for Flyio.ai — CMS, Blog, Trips, Knowledge Base, Web Scraper coordination, and LLM Workflow Orchestration.",
    "version": "1.0.0",
    "contact": {
      "name": "Flyio.ai Engineering"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Admin JWT token acquired via POST /api/admin/login"
      }
    },
    "schemas": {
...
      "LLMGenerateRequest": {
        "type": "object",
        "required": ["prompt"],
        "properties": {
          "prompt": { "type": "string", "example": "Create a 5-day travel itinerary for Paris, France" },
          "urls": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Optional specific source URLs to crawl if data is missing in vector database",
            "example": ["https://en.wikipedia.org/wiki/Paris"]
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "LLMGenerateResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "request_id": { "type": "string", "example": "req_8f7b2c1d4e0a" },
          "status": { "type": "string", "enum": ["data_found", "data_required", "generation_failed"], "example": "data_found" },
          "plan": { "type": "object", "description": "Synthesized AI travel plan object" },
          "results": { "type": "array", "items": { "type": "object" } },
          "message": { "type": "string" },
          "scraped_sources": { "type": "array", "items": { "type": "string" } },
          "execution_time_ms": { "type": "number", "example": 1450 }
        }
      },
      "LLMStoreRequest": {
        "type": "object",
        "properties": {
          "source_url": { "type": "string", "example": "https://en.wikipedia.org/wiki/Paris" },
          "all": { "type": "boolean", "example": false }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/admin/llm/generate": {
      "post": {
        "summary": "Generate Travel Plan with Automated LLM & Scraper Ingestion Orchestration",
        "description": "Orchestrates the entire pipeline: creates request_id, calls LLM Generate API, automatically triggers scraper + JSON conversion + Qdrant vector store if data_required, retries LLM plan generation, logs lifecycle events, and returns synthesized travel plan.",
        "tags": ["LLM Orchestration"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LLMGenerateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan generated successfully",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LLMGenerateResponse" } } }
          },
          "502": { "description": "LLM or Scraper Service communication error" }
        }
      }
    },
    "/api/admin/llm/store": {
      "post": {
        "summary": "Ingest Knowledge Base chunks into LLM Vector Store (Qdrant)",
        "tags": ["LLM Orchestration"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LLMStoreRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Documents ingested into vector database" }
        }
      }
    },
...
    "/api/admin/prompts": {
      "get": {
        "summary": "List prompt history",
        "tags": ["Prompt History"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Paginated prompt history" }
        }
      }
    },
    "/api/admin/prompts/{request_id}": {
      "get": {
        "summary": "Get prompt detail and cross-service request_events lifecycle trace",
        "tags": ["Prompt History"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "request_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Prompt detail and correlated events timeline" },
          "404": { "description": "Prompt record not found" }
        }
      },
      "delete": {
        "summary": "Delete prompt record and associated event traces",
        "tags": ["Prompt History"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "request_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deleted successfully" }
        }
      }
    },
...
  }
}